% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nztt-package.R
\docType{package}
\name{nztt-package}
\alias{nztt-package}
\alias{_PACKAGE}
\title{nztt: Fun with the New Zealand Soil Texture Triangle}
\description{
Tools to plot the soil texture triangle used in New Zealand, and to classify texture data into the NZ soil texture classes.
}
\author{
\strong{Maintainer}: Pierre Roudier \email{roudierp@landcaresearch.co.nz}

Other contributors:
\itemize{
  \item Lauren O'Brien \email{obrienl@landcaresearch.co.nz} [contributor]
  \item Kirstin Deuss \email{deussk@landcaresearch.co.nz} [contributor]
}

}
\keyword{internal}
