% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{nztt}
\alias{nztt}
\title{New Zealand Soil Texture Triangle}
\format{
`nztt`
A data frame with 42 rows and 5 columns:
\describe{
  \item{clay, sand, silt}{Percentages for clay, sand, and silt}
  \item{name}{Name of the texture classes}
  \item{code}{Code for the texture classes}
}
}
\source{
Milne, J. D. G., B. Clayden, P. L. Singleton, and A. D. Wilson. 1995. Soil Description Handbook. Revised. Lincoln, N.Z.: Manaaki Whenua Press.
}
\usage{
nztt
}
\description{
New Zealand Soil Texture Classification Data
}
\author{
Lauren O'Brien
}
\keyword{datasets}
