% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{palette_nzsc}
\alias{palette_nzsc}
\title{Colour Palette for the New Zealand Soil Classification Orders}
\format{
`palette_nzsc`
A vector of 15 hexadecimnal colours.
}
\usage{
palette_nzsc
}
\description{
Colour Palette for the New Zealand Soil Classification Orders. Colours were sourced from the S-Map portal.
}
\author{
Pierre Roudier
}
\keyword{datasets}
