% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{palette_nztt}
\alias{palette_nztt}
\title{Colour Palette for the New Zealand Soil Texture Triangle}
\format{
`palette_nztt`
A vector of 11 hexadecimnal colours.
}
\usage{
palette_nztt
}
\description{
Colour palette for the New Zealand Soil Texture Classification Data
}
\author{
Pierre Roudier, Lauren O'Brien
}
\keyword{datasets}
