% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{palette_smap}
\alias{palette_smap}
\title{Colour Palette for the S-Map Soil Texture Triangle}
\format{
`palette_nztt`
A vector of 4 hexadecimnal colours.
}
\usage{
palette_smap
}
\description{
Colour palette for the S-Map Soil Texture Classification Data
}
\author{
Pierre Roudier, Lauren O'Brien
}
\keyword{datasets}
