% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{smaptt}
\alias{smaptt}
\title{S-Map Soil Texture Triangle}
\format{
`smaptt`
A data frame with 15 rows and 5 columns:
\describe{
  \item{clay, sand, silt}{Percentages for clay, sand, and silt}
  \item{name}{Name of the texture classes}
  \item{code}{Code for the texture classes}
}
}
\usage{
smaptt
}
\description{
S-Map Soil Texture Classification Data
}
\author{
Pierre Roudier
}
\keyword{datasets}
