% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texture_class.R
\name{texture_class}
\alias{texture_class}
\title{Get Texture Class from Texture Data}
\usage{
texture_class(clay, sand, silt)
}
\arguments{
\item{clay}{.}

\item{sand}{.}

\item{silt}{.}
}
\value{
description .
}
\description{
texture_class
}
\examples{

texture_class(clay = 40, silt = 14, sand = 46)


# Load example data
data(wairau)

# Calculate texture classes
tx <- texture_class(clay = wairau$clay, sand = wairau$sand, silt = wairau$silt)
print(tx)

# Check the results
cbind(tx$name, wairau$name)

# Example using the tidyverse verbs

library(dplyr)
data(wairau)
wairau \%>\%
  select(id, clay, sand, silt) \%>\%
  mutate(
    texture = texture_class(clay = clay, sand = sand, silt = silt)$name
  )

}
\author{
Pierre Roudier
}
