% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{wairau}
\alias{wairau}
\title{Soil Texture Data from the Wairau Valley}
\format{
`wairau`
A data frame with 10 rows and 6 columns:
\describe{
  \item{id}{A unique sample identifier}
  \item{clay, sand, silt}{Percentages for clay, sand, and silt}
  \item{name}{Name of the texture classes}
  \item{code}{Code for the texture classes}
}
}
\source{
Milne, J. D. G., B. Clayden, P. L. Singleton, and A. D. Wilson. 1995. Soil Description Handbook. Revised. Lincoln, N.Z.: Manaaki Whenua Press.
}
\usage{
wairau
}
\description{
Soil Texture Data from the Wairau Valley
}
\author{
Kirstin Deuss
}
\keyword{datasets}
