% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texture_class.R
\name{texture_class}
\alias{texture_class}
\title{Get Texture Class from Texture Data}
\usage{
texture_class(clay, sand, silt, triangle = "nz", rescale_psd = FALSE)
}
\arguments{
\item{clay}{.}

\item{sand}{.}

\item{silt}{.}

\item{triangle}{Which texture triangle to use. Currently implemented are the NZ texture triangle (option "nz", default) or the S-Map texture triangle (option "smap").}

\item{rescale_psd}{Does the PSD data needs to be rescaled from [0, 1] to [0, 100]? Defaults to `FALSE`.}
}
\value{
description .
}
\description{
texture_class
}
\examples{

texture_class(clay = 40, silt = 14, sand = 46)
texture_class(clay = 40, silt = 14, sand = 46, triangle = "smap")

# Load example data
data(wairau)

# Calculate texture classes
tx <- texture_class(clay = wairau$clay, sand = wairau$sand, silt = wairau$silt)
print(tx)

# Check the results
cbind(tx$name, wairau$name)

# Example using the tidyverse verbs

library(dplyr)
data(wairau)
wairau \%>\%
  select(id, clay, sand, silt) \%>\%
  mutate(
    texture = texture_class(clay = clay, sand = sand, silt = silt)$name
  )

}
\author{
Pierre Roudier
}
