% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texture_class.R
\name{texture_class}
\alias{texture_class}
\title{Get Texture Class from Texture Data}
\usage{
texture_class(clay, sand, silt)
}
\arguments{
\item{clay}{.}

\item{sand}{.}

\item{silt}{.}
}
\value{
description .
}
\description{
texture_class
}
\examples{

texture_class(clay = 40, silt = 14, sand = 46)

}
\author{
Pierre Roudier
}
